input: misc: et51x: Reset hardware once, not after powering up.
The original driver doesn't turn the device off to save power while it's
not in use, but careful testing reveals that is properly possible.

Resetting the device after powering it causes all kinds of inexplicable
side-effects, such as the dreaded error -59 on initial enroll, and the
TZ-app requesting another hardware reset when starting an auth or enroll
session.

Removing this reset fixes all aforementioned issues. The reset has to be
done once, when the device is initially powered up.

Change-Id: I432c7f01d00f3e3be26ce6610b504d7d6afcd21d
Signed-off-by: MarijnS95 <marijns95@gmail.com>