input: misc: et51x: Do not toggle power multiple times during probe.
Power has to stay enabled until the TZ initializes, or it'll request a
device reset on every finger detection operation. Different combinations
of powering the device on and resetting the device when the HAL
initializes have no effect on this requested reset.
This might also be the source for the yet inexplicable error -59 on
initial enroll.

It is also impossible for the driver to turn the power off before the TZ
has been initialized. If this happens, no errors are reported but the
IRQ is not triggered when a finger touches the sensor.

Change-Id: Ieaad4da024c39f4469695b3475d56d673cae4be0
Signed-off-by: MarijnS95 <marijns95@gmail.com>