input: misc: et51x: Drop regulator load and change to block comments.
The original driver sets the load to 0, which is a no-op since it
doesn't contribute to the total load the regulator has to supply. This
is a way of saying that the driver developers didn't know what the
sensor is using. Since there are no public product sheets available,
we don't know this either.

Change-Id: I5c51ab2929151270869f9006d71750db97b09a71
Signed-off-by: MarijnS95 <marijns95@gmail.com>