input: misc: et51x: Switch vdd_ana regulator to 3.3 volts.
This regulator has always been on 3.3 volts on the stock kernel for both
Nile and Ganges. Nile uses it temporarily before switching to a lower
1.8v when an FPC sensor is detected, leading to believe that it's safe
to use this.

Ganges needs this higher voltage or the sensor becomes highly unstable
and unresponsive - something the Nile Egistec sensor didn't suffer from.

As a safety measure, initialize the regulator to FPC voltage (1.8) while
checking which sensor is installed, before scaling up to 3.3v.

Signed-off-by: Marijn Suijten <marijns95@gmail.com>