input: ts: sec_ts: Enable Grip Rejection and disable SideTouch at init
The sidetouch functionality is not handled in Open Devices builds,
so it's wise to disable it in order to avoid any possible mishandling
of touch events.

Also, to improve the user experience, keep the Grip Rejection feature
always enabled, as the device is expected to be used in portrait mode
for most of the time: this will not compromise functionality while
the device is being used in landscape mode in any way.