input: ts: sec_ts: Fix touchscreen init in DrIC not ready case
When the userspace is slow, it will try to initialize the
touchscreen before getting the display DriverIC ready to rock:
this will result in a bad fault because our TS is integrated
in the DrIC and needs it to be up and running in order to be
able to give us good replies.

Add a mechanism to delay the touchscreen initialization in time
until we expect the userspace to be almost ready.
If the userspace is not ready, keep delaying the initialization
for one second everytime we fail, until we succeed.