media: msm: camera_v2: Pass mandatory vaddr to dma_buf_vunmap.
The comment specifies that this argument is ignored while the
implementation definitely BUGs if it is not equal to ->vmap_ptr.

The comparison against vmap_ptr is there for consistency. To stick to
that the inverse of msm_buf_mngr_handle_cont_cmd is applied: all buffers
are removed and freed from cont_qhead in reverse, and the last entry
with matching paddr executes the call to dma_buf_vunmap.

Finally, the "somewhat hidden" loop variable cnt--; at the bottom is
pulled up to make the iteration more obvious, and the comparison of bool
unmap against integer 1 is removed.

Fixes: 3b4a7a3864bd586a10f0b13737f7909532b2ddad
Signed-off-by: Marijn Suijten <marijns95@gmail.com>