leds: qti-tri-led: Write known value to TRILED HW at init time
The bootloader may leave one (or more) of the LEDs up when booting
the HLOS: in that case, the LED will be stuck because we are
disallowing to write brightness when br_new==br_old.

We initialize the brightness to zero when initializing the driver,
so also initialize the hardware to known values: this way, the
driver init matches with the hardware status.