input: ts: sec_ts: Read DT mandatory touch-rst-gpio and expected DevID
Add a mandatory DT property "somc,touch-rst-gpio" and also read
the expected device id "somc,expected-device-id" array (or fall
back to Kumano platform defaults for the latter).

The touch reset gpio is mandatory to avoid hardware damage on
some platforms when the device id does not match (when forcing
a ts reset in this driver).