arm64: DT: Seine-touch: Declare i2c-touch-active property
Qualcomm now wants us to specify this property because, for some
reason, they want to compile in all touchscreen drivers on Earth
and ignore the probe if a property that is external to the touch
driver node does not correspond to the touch driver node itself.

All because enabling in DT only what's needed is old-school.

P.S.: Are you serious?